GNU => MIT good thoughts on this.
As a moderate license noob, I didnt know which way to go.  My heart says open source, but I dont want to be so restrictive WITH my license that strangers cannot fork and edit and repost as they please.  Not my job to care what people do with this, except to blackmail me.